Separating other_views.py to files
Moved functions from other_views.py to comment_views.py and followers.py. Changed __init__.py for views.
